# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Library API",
        "contact": {
            "email": "bmpcorp9@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/artist": {
            "get": {
                "tags": [
                    "Artists",
                    "List"
                ],
                "summary": "Get a paginated list of artists",
                "operationId": "2e5823558d57359cd1f4661d64590ef8",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query for artist names or music genres",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "no_explicit",
                        "in": "query",
                        "description": "Filter out artists with explicit content (1 for true, 0 for false)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of artists retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Artist Name"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "This Band Rocks!"
                                                            },
                                                            "genres": {
                                                                "type": "string",
                                                                "example": "Pop, Rock"
                                                            },
                                                            "country_id": {
                                                                "type": "integer",
                                                                "example": 15
                                                            },
                                                            "logo": {
                                                                "type": "string",
                                                                "example": "https://example.com/logo.png"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 100
                                                        },
                                                        "page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "pages": {
                                                            "type": "integer",
                                                            "example": 10
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/artist/{id}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Get a single artist by ID",
                "operationId": "2354ab7223cc9f78ad6d094ccfdc5662",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the artist to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "item": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Artist Name"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "This Band Rocks!"
                                                        },
                                                        "genres": {
                                                            "type": "string",
                                                            "example": "Pop, Rock"
                                                        },
                                                        "logo": {
                                                            "type": "string",
                                                            "example": "https://example.com/logo.png"
                                                        },
                                                        "favorite_of_users_count": {
                                                            "type": "integer",
                                                            "example": 12
                                                        },
                                                        "albums": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 101
                                                                    },
                                                                    "title": {
                                                                        "type": "string",
                                                                        "example": "Album Title"
                                                                    },
                                                                    "description": {
                                                                        "type": "string",
                                                                        "example": "First album after reunition"
                                                                    },
                                                                    "genres": {
                                                                        "type": "string",
                                                                        "example": "Pop, Rock"
                                                                    },
                                                                    "year": {
                                                                        "type": "integer",
                                                                        "example": 2023
                                                                    },
                                                                    "song_count": {
                                                                        "type": "integer",
                                                                        "example": 9
                                                                    },
                                                                    "has_explicit_lyrics": {
                                                                        "type": "boolean",
                                                                        "example": true
                                                                    },
                                                                    "cover": {
                                                                        "type": "string",
                                                                        "example": "https://example.com/album_cover.png"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "country": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "code": {
                                                                    "type": "string",
                                                                    "example": "RU"
                                                                },
                                                                "title": {
                                                                    "type": "string",
                                                                    "example": "Russia"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/artist/{id}/favorite": {
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Toggle an artist as favorite for the authenticated user",
                "operationId": "9183a3a0faa4ccee52766cb299460fd3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the artist to toggle as favorite",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist favorite status toggled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Artist added to favorites"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artist not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}